order-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
aiokafka==0.10.0
asyncpg==0.29.0
python-dotenv==1.0.1